Group account logic on accounts folder
