+++ packages/common/src/accounts/constants.ts
@@ -1,0 +1,5 @@
+export const ACCOUNTS_DIR_NAME = 'accounts';
+
+export enum ACCOUNT_TYPES {
+    LOCAL = 'LOCAL',
+}
