+++ packages/common/src/constants.ts
@@ -1,6 +1,1 @@
-export enum ACCOUNT_TYPES {
-    LOCAL = 'LOCAL',
-}
-
-export const ACCOUNTS_DIR_NAME = 'accounts';
 export const JSON_FILE_EXTENSION = '.json';
