+++ packages/common/src/utils/index.ts
@@ -1,2 +1,1 @@
 export {arrayHasItems} from './array-has-items';
-export {createAccountInstance} from './create-account-instance';
